requirements.txt
chardet==5.1.0
datasketch==1.5.8
joblib==1.2.0
numpy==1.23.5
pandas==1.5.2
python_dateutil==2.8.2
pytorch-lightning==1.8.6
scikit-learn==1.2.0
scipy==1.9.3
# torch==1.13.1
torchmetrics==0.11.0
transformers==4.25.1
xxhash==3.2.0
